Move FlagSet interface to a common place.
The two copies the FlagSet interface are identical and are required to
be so.  It seems better to have them in place.  We put this in a public
directory as the interface defined in driver/driver.go is already
public.
